// src/app/core/models/perio-maintenance.model.ts
export interface CreatePerioMaintenanceDto {
  hygieneProvider: string;
  periodonitst: string;
  cc?: string;
  refferingDoctor?: string;
  HhReviewedNoCharges?: boolean;
  HhReviewedUpdated?: boolean;
  parq?: boolean;
  bp?: string;
  wnl?: boolean;
  findings?: string;
  stageI?: boolean;
  stageII?: boolean;
  stageIII?: boolean;
  stageIV?: boolean;
  gradeA?: boolean;
  gradeB?: boolean;
  gradeC?: boolean;
  noExam?: boolean;
  examType?: string;
  examNotes?: string;
  perioChartUpdate?: boolean;
  ultrasonic?: boolean;
  handInstrumentation?: boolean;
  implantInstrument?: boolean;
  sensitivityNote?: string;
  oralSedation?: string;
  n202?: string;
  informedConsentTxt?: boolean;
  localAnesthetic?: string;
  amt?: string;
  injections?: string;
  informedConsent?: boolean;
  fmx?: boolean;
  bwx?: boolean;
  stable?: boolean;
  unstable?: boolean;
  patientStatusNote?: string;
  oralHygiene?: string;
  electricTb?: boolean;
  floss?: boolean;
  flRx?: boolean;
  endTuft?: boolean;
  manualTb?: boolean;
  proxyBrush?: boolean;
  rinse?: boolean;
  softPicks?: boolean;
  homeCareNote?: string;
  supraPlaqueGen?: boolean;
  supraPlaqueLoc?: boolean;
  supraPlaqueS?: boolean;
  supraPlaqueM?: boolean;
  supraPlaqueH?: boolean;
  flossGen?: boolean;
  flossLoc?: boolean;
  flossS?: boolean;
  flossM?: boolean;
  flossH?: boolean;
  subPlaqueGen?: boolean;
  subPlaqueLoc?: boolean;
  subPlaqueS?: boolean;
  subPlaqueM?: boolean;
  subPlaqueH?: boolean;
  subCalculasGen?: boolean;
  subCalculasLoc?: boolean;
  subCalculasS?: boolean;
  subCalculasM?: boolean;
  subCalculasH?: boolean;
  inflammationGen?: boolean;
  inflammationLoc?: boolean;
  inflammationS?: boolean;
  inflammationM?: boolean;
  inflammationH?: boolean;
  bleedingGen?: boolean;
  bleedingLoc?: boolean;
  bleedingS?: boolean;
  bleedingM?: boolean;
  bleedingH?: boolean;
  healthy?: boolean;
  stipled?: boolean;
  rolled?: boolean;
  spongy?: boolean;
  pink?: boolean;
  smooth?: boolean;
  red?: boolean;
  fibrotic?: boolean;
  presentsWithNote?: string;
}